Upgraded SAML2 Handler to OpenSAML V4. Java 11. Sling 12. Reviewed and optimized embedded OpenSAML dependencies and updated versions.
update docs

made encryption and signing an optional configuration

Backed out SLO for now and moved that to a different branch. Logout in this branch simply drops the credentials for this Service Provider and redirects the client to the URL configured

Added error handling runtime exception if the userID is not configured properly. JCR Group ID's now seem to slash in front so added condition to allow that. Updated realm-sling-export.json (for Keycloak demo) adding attribute mappers for client, http://localhost:8080

updated readme and added localhost test package